\name{FredsStatsPack-package}
\alias{FredsStatsPack-package}
\alias{FredsStatsPack}
\docType{package}
\title{
\packageTitle{FredsStatsPack}
}
\description{
\packageDescription{FredsStatsPack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FredsStatsPack}
\packageIndices{FredsStatsPack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FredsStatsPack}

Maintainer: \packageMaintainer{FredsStatsPack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
